Remove redundant UI code and simplify frontend modules
