+++ docs/directory.md
@@ -188,7 +188,7 @@ The frontend loads scripts in a strict dependency order via `ui/index.html`:
 26. `chat-window.js` — verified Chat window, ownership/recovery coordination, and dedicated display bootstrap (`window.LlamaGui.chatWindow`)
 27. `benchmark-ui.js` — Benchmarking tab controls, argument adapter, output polling, and session-only summaries (`window.LlamaGui.benchmarkUi`)
 28. `inference-stats.js` — pure metrics/slots normalization and target-keyed inference snapshot engine (`window.LlamaGui.inferenceStats`)
-29. `monitor-ui.js` — Monitor tab system/GPU polling, process-output terminal, shared inference snapshot rendering, card visibility preferences (`window.LlamaGui.monitorUi`)
+29. `monitor/*` package, loaded in order: `monitor-internal.js` (private dependency links), `monitor-dom.js` (formatting/DOM primitives), `monitor-preferences.js` (visibility/order/drag), `monitor-system.js` (CPU/RAM/disk and sample presentation), `monitor-gpu.js` (GPU/setup/state cards), `monitor-terminal.js` (runtime/header and terminal), `monitor-inference.js` (shared snapshot rendering), `monitor-polling.js` (system telemetry lifecycle), `monitor-main.js` (`window.LlamaGui.monitorUi` facade and initialization)
 30. `shell-ui.js` — grouped navigation, responsive navigation drawer, and the shared sidebar runtime summary (`window.LlamaGui.shellUi`)
 31. `app.js` — main orchestration (wires everything together)
 
@@ -221,6 +221,13 @@ retains polling, target reconciliation, and epoch invalidation; Monitor renders
 the emitted snapshot in both views. A hidden host keeps inference polling while
 detached Chat is open, while system telemetry keeps its panel/document gates.
 Detached Chat consumes host snapshots and creates no inference engine or poller.
+
+Monitor follows the same concern-owned closure pattern as Manager. Its internal
+links are private to `ui/js/monitor/`; loading and configuration bind no controls
+and start no polling. System telemetry owns its accepted sample and exposes only
+the sample interval needed by system-card labels. Preferences retain the newest
+sample during a drag and flush it once when the drag ends. GPU reconciliation
+then reapplies visibility and order without rebuilding unchanged nodes.
 
 ### Frontend Module Reference
 
@@ -288,7 +295,15 @@ Detached Chat consumes host snapshots and creates no inference engine or poller.
 | `ui/js/chat-window.js` | `window.LlamaGui.chatWindow` | Dedicated Chat display, verified main/popup bridge, exclusive workspace ownership, handoff, and recovery checkpoints |
 | `ui/js/benchmark-ui.js` | `window.LlamaGui.benchmarkUi` | Benchmarking tab source selection, benchmark-specific controls, compatible argument building for `llama-bench`/`llama-perplexity`, readiness/status badges, process actions, output polling, and session-only summaries |
 | `ui/js/inference-stats.js` | `window.LlamaGui.inferenceStats` | DOM-free `parseMetricsText`, `normalizeSlots`, and `createInferenceStats`; each engine owns target-keyed baselines, rate samples, sequence, and source availability, emitting data snapshots through its callback |
-| `ui/js/monitor-ui.js` | `window.LlamaGui.monitorUi` | Monitor tab: system-stats polling with visibility gating and truthful status badge, process-output terminal (always-follow output, trim, cursor-preserving clear), dynamically reconciled GPU cards in the shared metrics grid, setup/state rendering with backend-supplied platform guidance, hidden-card preferences with tolerant persistence, and rendering the shared inference snapshot in the fixed stats bar and the Inference card; retains compatibility aliases for the three inference-core methods |
+| `ui/js/monitor/monitor-internal.js` | `window.LlamaGui._monitorInternal` (private) | Live dependencies and named concern links; no shared domain-state bag |
+| `ui/js/monitor/monitor-dom.js` | private `dom` methods | Monitor formatting and shared DOM primitives, including in-place text and progress meters |
+| `ui/js/monitor/monitor-preferences.js` | private `preferences` methods | Hidden-card storage, order, restore focus, keyboard/drag controls, and the latest deferred telemetry sample; index-fallback identities remain session-only |
+| `ui/js/monitor/monitor-system.js` | private `system` methods | CPU/RAM/disk readings and accepted sample presentation; waits for card dragging to finish before rendering |
+| `ui/js/monitor/monitor-gpu.js` | private `gpu` methods | Stable GPU identities, in-place card reconciliation, vendor state/setup guidance, and preference application |
+| `ui/js/monitor/monitor-terminal.js` | private `terminal` methods | Runtime/header presentation and always-follow, bounded terminal output; lifecycle, input visibility and output cursors stay in `app.js` |
+| `ui/js/monitor/monitor-inference.js` | private `inference` methods | Monitor inference availability/card and fixed stats bar, including a supplied detached document; consumes shared snapshots |
+| `ui/js/monitor/monitor-polling.js` | private `polling` methods | System telemetry visibility gates, timer/controller/generation, last accepted sample, Recheck, and live badge |
+| `ui/js/monitor/monitor-main.js` | `window.LlamaGui.monitorUi` | Dependency configuration, initialization, stable public facade and the retained three inference-core compatibility aliases |
 | `ui/js/app.js` | `window.LlamaGui` (global) | Main UI orchestration. Manages tab switching, server launch/stop, output polling, the single inference poll cycle that feeds one shared snapshot to the fixed bar and Monitor, toasts, module initialization, and cache-busting reload |
 | `ui/css/style.css` | — | Shared page headings/action bars, controls, surfaces, and responsive layout. Contains no color literals and no `[data-theme=…]` selectors — all color lives in `ui/css/tokens.css` |
 | `ui/js/shell-ui.js` | `window.LlamaGui.shellUi` | Workspace/maintenance navigation, current-page semantics, mobile drawer focus and dismissal, and the shared sidebar summary of the authoritative local runtime or registered external server |
@@ -1211,5 +1226,5 @@ Prefer `rg` for local search. On Windows/PowerShell, use patterns like `rg -n "p
 | `docs/upstream-changes.md` | llama.cpp upstream changes needing coordinated GUI updates |
 | `docs/software-versioning-policy.md` | CalVer versioning and stable-release policy |
 | `docs/frontend-module-split-plan.md` | Completed Tier-1 frontend module-split recipe and implementation record |
-| `docs/frontend-maintainability-tier-2-plan.md` | Tier-2 frontend maintainability plan in progress: Sessions 0–5 complete, Session 6 next; module boundaries, implementation order, and verification gates |
+| `docs/frontend-maintainability-tier-2-plan.md` | Tier-2 frontend maintainability plan in progress: Sessions 0–6 complete, Session 7 next; module boundaries, implementation order, and verification gates |
 | `docs/images/` | Screenshots used by README.md |
