+++ package.json
@@ -3,7 +3,7 @@
   "private": true,
   "scripts": {
     "test": "npm run test:unit && npm run test:flags && npm run test:frontend",
-    "test:unit": "npm run test:syntax && node tests/frontend/custom_launch_args_unit.cjs && node tests/frontend/launch_args_unit.cjs && node tests/frontend/chat_template_selection_unit.cjs && node tests/frontend/output_cursor_unit.cjs && node tests/frontend/inference_stats_unit.cjs && node tests/frontend/monitor_ui_unit.cjs && node tests/frontend/process_lifecycle_unit.cjs && node tests/frontend/benchmark_args_unit.cjs && node tests/frontend/chat_rendering_unit.cjs && node tests/frontend/chat_ui_unit.cjs && node tests/frontend/chat_window_unit.cjs && node tests/frontend/chat_tools_unit.cjs && node tests/frontend/chat_compaction_unit.cjs && node tests/frontend/character_cards_unit.cjs && node tests/frontend/sampler_presets_unit.cjs && node tests/frontend/presets_unit.cjs && node tests/frontend/preset_roving_focus_unit.cjs && node tests/frontend/model_switch_ui_unit.cjs && node tests/frontend/hf_download_ui_unit.cjs && node tests/frontend/remote_tunnel_ui_unit.cjs && node tests/frontend/api_tab_unit.cjs && node tests/frontend/external_server_ui_unit.cjs && node tests/frontend/shared_services_unit.cjs && node tests/frontend/manager_releases_unit.cjs && node tests/frontend/manager_lifecycle_unit.cjs && node tests/frontend/manager_model_cache_unit.cjs && node tests/frontend/manager_model_dir_unit.cjs && node tests/frontend/theme_ui_unit.cjs && npm run test:frontend:modules && npm run test:flag-definitions",
+    "test:unit": "npm run test:syntax && node tests/frontend/custom_launch_args_unit.cjs && node tests/frontend/launch_args_unit.cjs && node tests/frontend/chat_template_selection_unit.cjs && node tests/frontend/output_cursor_unit.cjs && node tests/frontend/inference_stats_unit.cjs && node --test tests/frontend/monitor_package_unit.cjs tests/frontend/monitor_system_cards_unit.cjs tests/frontend/monitor_gpu_cards_unit.cjs tests/frontend/monitor_inference_rendering_unit.cjs tests/frontend/monitor_terminal_runtime_unit.cjs tests/frontend/monitor_polling_unit.cjs tests/frontend/monitor_card_preferences_unit.cjs && node tests/frontend/process_lifecycle_unit.cjs && node tests/frontend/benchmark_args_unit.cjs && node tests/frontend/chat_rendering_unit.cjs && node tests/frontend/chat_ui_unit.cjs && node tests/frontend/chat_window_unit.cjs && node tests/frontend/chat_tools_unit.cjs && node tests/frontend/chat_compaction_unit.cjs && node tests/frontend/character_cards_unit.cjs && node tests/frontend/sampler_presets_unit.cjs && node tests/frontend/presets_unit.cjs && node tests/frontend/preset_roving_focus_unit.cjs && node tests/frontend/model_switch_ui_unit.cjs && node tests/frontend/hf_download_ui_unit.cjs && node tests/frontend/remote_tunnel_ui_unit.cjs && node tests/frontend/api_tab_unit.cjs && node tests/frontend/external_server_ui_unit.cjs && node tests/frontend/shared_services_unit.cjs && node tests/frontend/manager_releases_unit.cjs && node tests/frontend/manager_lifecycle_unit.cjs && node tests/frontend/manager_model_cache_unit.cjs && node tests/frontend/manager_model_dir_unit.cjs && node tests/frontend/theme_ui_unit.cjs && npm run test:frontend:modules && npm run test:flag-definitions",
     "test:syntax": "node tests/frontend/js_syntax_check.cjs",
     "test:frontend:modules": "node tests/frontend/module_namespace_unit.cjs",
     "test:flag-definitions": "node tests/frontend/flag_definitions_unit.cjs",
