+++ ui/index.html
@@ -1902,7 +1902,15 @@
 <script src="/js/chat-window.js?v=chat-window-1"></script>
 <script src="/js/benchmark-ui.js?v=revamp-2"></script>
 <script src="/js/inference-stats.js?v=1"></script>
-<script src="/js/monitor-ui.js?v=monitor-1"></script>
+<script src="/js/monitor/monitor-internal.js?v=1"></script>
+<script src="/js/monitor/monitor-dom.js?v=1"></script>
+<script src="/js/monitor/monitor-preferences.js?v=1"></script>
+<script src="/js/monitor/monitor-system.js?v=1"></script>
+<script src="/js/monitor/monitor-gpu.js?v=1"></script>
+<script src="/js/monitor/monitor-terminal.js?v=1"></script>
+<script src="/js/monitor/monitor-inference.js?v=1"></script>
+<script src="/js/monitor/monitor-polling.js?v=1"></script>
+<script src="/js/monitor/monitor-main.js?v=1"></script>
 <script src="/js/shell-ui.js?v=shell-1"></script>
 <script src="/js/app.js?v=revamp-4"></script>
 </body>
