--- ui/js/monitor-ui.js
@@ -1,1689 +1,0 @@
-// Monitor tab UI: system/GPU polling, process-output terminal, shared
-// inference snapshot rendering, and card visibility preferences.
-//
-// Ownership rules:
-// - Process lifecycle, launch/stop, and cursor ownership stay in app.js;
-//   this module only renders output lines and clears the terminal.
-// - Inference telemetry is fetched by app.js's existing stats poller; this
-//   module renders the shared snapshots produced by inference-stats.js.
-(function () {
-    "use strict";
-
-    const root = window.LlamaGui = window.LlamaGui || {};
-
-    // ════════════════════════════════════════════════════════════════════
-    // Pure formatting helpers
-    // ════════════════════════════════════════════════════════════════════
-
-    const BYTE_UNITS = ["B", "KB", "MB", "GB", "TB", "PB"];
-
-    function finiteOrNull(value) {
-        if (value === null || value === undefined) return null;
-        const num = Number(value);
-        return Number.isFinite(num) ? num : null;
-    }
-
-    function finiteNonNegativeOrNull(value) {
-        const num = finiteOrNull(value);
-        return num !== null && num >= 0 ? num : null;
-    }
-
-    function clampPercent(value) {
-        const num = finiteOrNull(value);
-        if (num === null) return null;
-        return Math.max(0, Math.min(100, num));
-    }
-
-    function formatBytes(bytes) {
-        const value = finiteNonNegativeOrNull(bytes);
-        if (value === null) return "Not available";
-        let scaled = value;
-        let unitIndex = 0;
-        while (scaled >= 1024 && unitIndex < BYTE_UNITS.length - 1) {
-            scaled /= 1024;
-            unitIndex += 1;
-        }
-        const digits = unitIndex === 0 ? 0 : (scaled >= 100 ? 0 : 1);
-        return `${scaled.toFixed(digits)} ${BYTE_UNITS[unitIndex]}`;
-    }
-
-    function formatRate(bytesPerSecond) {
-        const value = finiteNonNegativeOrNull(bytesPerSecond);
-        if (value === null) return "Not available";
-        return `${formatBytes(value)}/s`;
-    }
-
-    function formatPercentValue(value, digits = 0) {
-        const clamped = clampPercent(value);
-        if (clamped === null) return "Not available";
-        return clamped.toFixed(digits);
-    }
-
-    function formatTokens(value) {
-        const num = finiteNonNegativeOrNull(value);
-        if (num === null) return "--";
-        return Math.round(num).toLocaleString();
-    }
-
-    function formatSpeed(tokensPerSecond) {
-        const num = finiteNonNegativeOrNull(tokensPerSecond);
-        if (num === null) return "--";
-        return num.toFixed(1);
-    }
-
-    function formatClock(epochSeconds) {
-        const num = finiteOrNull(epochSeconds);
-        if (num === null) return "--:--:--";
-        const date = new Date(num * 1000);
-        const pad = (part) => String(part).padStart(2, "0");
-        return `${pad(date.getHours())}:${pad(date.getMinutes())}:${pad(date.getSeconds())}`;
-    }
-
-    function shortGpuId(id) {
-        const text = String(id || "");
-        if (text.length <= 24) return text;
-        return `${text.slice(0, 12)}\u2026${text.slice(-5)}`;
-    }
-
-    // ════════════════════════════════════════════════════════════════════
-    // Hidden-card preference store (view-only; telemetry is unaffected)
-    // ════════════════════════════════════════════════════════════════════
-
-    const HIDDEN_STORAGE_KEY = "llama_gui_monitor_hidden_cards";
-    const HIDDEN_MAX_ENTRIES = 100;
-    const HIDDEN_MAX_KEY_LENGTH = 256;
-    const HIDDEN_MAX_LABEL_LENGTH = 120;
-
-    const ORDER_STORAGE_KEY = "llama_gui_monitor_card_order";
-    const ORDER_MAX_ENTRIES = 100;
-    const ORDER_MAX_KEY_LENGTH = 256;
-    // Cards reorder within their own container flow only (metrics grid, GPU
-    // grid, state cards, setup cards). TODO: upgrade path is a single
-    // flattened grid if cross-container moves are ever wanted.
-    const ORDER_CONTAINER_SELECTOR = ".monitor-drag-container";
-    // Chromium suppresses clicks and text selection under a draggable
-    // ancestor, so the dedicated grip is the only draggable element.
-    const ORDER_DRAG_HANDLE_SELECTOR = ".monitor-drag-handle";
-
-    function isSessionOnlyKey(key) {
-        // Index-fallback GPU identities are not stable across boots, so their
-        // hides last for the session only.
-        return typeof key === "string" && key.startsWith("gpu:") && key.includes(":index:");
-    }
-
-    function normalizeHiddenEntries(value) {
-        if (!Array.isArray(value)) return [];
-        const seen = new Set();
-        const entries = [];
-        for (const entry of value) {
-            if (entries.length >= HIDDEN_MAX_ENTRIES) break;
-            if (!entry || typeof entry !== "object") continue;
-            const key = typeof entry.key === "string" ? entry.key : "";
-            const label = typeof entry.label === "string" ? entry.label : "";
-            if (!key || key.length > HIDDEN_MAX_KEY_LENGTH || seen.has(key)) continue;
-            seen.add(key);
-            entries.push({ key, label: label.slice(0, HIDDEN_MAX_LABEL_LENGTH) });
-        }
-        return entries;
-    }
-
-    function normalizeOrderEntries(value) {
-        if (!Array.isArray(value)) return [];
-        const seen = new Set();
-        const order = [];
-        for (const key of value) {
-            if (order.length >= ORDER_MAX_ENTRIES) break;
-            if (typeof key !== "string" || !key || key.length > ORDER_MAX_KEY_LENGTH) continue;
-            if (seen.has(key)) continue;
-            seen.add(key);
-            order.push(key);
-        }
-        return order;
-    }
-
-    // ════════════════════════════════════════════════════════════════════
-    // Module state
-    // ════════════════════════════════════════════════════════════════════
-
-    const POLL_INTERVAL_MS = 2000;
-    const TERMINAL_MAX_LINES = 5000;
-    const TERMINAL_TRIM = 1000;
-
-    let deps = {};
-    let panelVisible = false;
-    let documentVisible = true;
-    let pollTimer = null;
-    let pollController = null;
-    let pollGeneration = 0;
-    let lastSample = null;
-    let lastSuccessAt = 0;
-    let lastPollFailed = false;
-    let everSucceeded = false;
-    // key -> { label, sessionOnly }
-    let hiddenCards = new Map();
-    let hiddenRestoreSignature = "";
-    // Effective card order (keys may exist across all containers); session-only
-    // index-fallback keys may ride along in memory but are never persisted.
-    let cardOrder = [];
-    let dragKey = null;
-    let dragContainer = null;
-    let dragActive = false;
-    // A sample that landed mid-drag; the GPU rebuild would destroy the dragged
-    // node and cancel the drag, so it is deferred until the drag ends.
-    let deferredRender = null;
-
-    function byId(id) {
-        return document.getElementById(id);
-    }
-
-    function requireDep(name) {
-        const value = deps[name];
-        if (typeof value !== "function") {
-            throw new Error(`Monitor UI dependency missing: ${name}`);
-        }
-        return value;
-    }
-
-    function configure(nextDeps) {
-        deps = Object.assign({}, deps, nextDeps || {});
-    }
-
-    // ════════════════════════════════════════════════════════════════════
-    // Live-status badge
-    // ════════════════════════════════════════════════════════════════════
-
-    function currentStatusState() {
-        if (!panelVisible || !documentVisible) return "paused";
-        if (!everSucceeded) return "unavailable";
-        if (pollController) return "refreshing";
-        if (lastPollFailed) return "stale";
-        return "live";
-    }
-
-    function renderLiveBadge() {
-        const badge = byId("monitor-live-badge");
-        const updated = byId("monitor-last-updated");
-        if (updated) updated.textContent = formatClock(lastSuccessAt || null);
-        if (!badge) return;
-        const state = currentStatusState();
-        const labels = {
-            live: "Live \u00b7 ~2 s",
-            refreshing: "Refreshing\u2026",
-            stale: "Stale \u00b7 retrying",
-            unavailable: "Unavailable",
-            paused: "Paused",
-        };
-        badge.textContent = `System telemetry · ${labels[state]}`;
-        badge.classList.toggle("badge-green", state === "live");
-        badge.classList.toggle("badge-neutral", state === "refreshing" || state === "paused");
-        badge.classList.toggle("badge-yellow", state === "stale");
-        badge.classList.toggle("badge-dim", state === "unavailable");
-    }
-
-    // ════════════════════════════════════════════════════════════════════
-    // System/GPU polling
-    // ════════════════════════════════════════════════════════════════════
-
-    function pollingShouldRun() {
-        return panelVisible && documentVisible;
-    }
-
-    function cancelScheduledPoll() {
-        if (pollTimer) {
-            clearTimeout(pollTimer);
-            pollTimer = null;
-        }
-    }
-
-    function abortInFlight() {
-        if (pollController) {
-            pollController.abort();
-            pollController = null;
-        }
-    }
-
-    function scheduleNextPoll() {
-        cancelScheduledPoll();
-        if (!pollingShouldRun()) return;
-        const interval = Number(deps.pollIntervalMs);
-        const delay = Number.isFinite(interval) && interval > 0 ? interval : POLL_INTERVAL_MS;
-        pollTimer = setTimeout(() => {
-            pollTimer = null;
-            pollSystemStats(false);
-        }, delay);
-    }
-
-    async function pollSystemStats(forceRefresh) {
-        if (pollController) pollController.abort();
-        const generation = ++pollGeneration;
-        const controller = new AbortController();
-        pollController = controller;
-        if (forceRefresh) renderLiveBadge();
-        try {
-            const fetchJson = requireDep("fetchJson");
-            const url = forceRefresh ? "/api/system-stats?refresh=1" : "/api/system-stats";
-            const data = await fetchJson(url, { signal: controller.signal });
-            if (generation !== pollGeneration) return;
-            lastSample = data;
-            lastSuccessAt = Number(data && data.sampled_at) || Date.now() / 1000;
-            lastPollFailed = false;
-            everSucceeded = true;
-            renderSample(data);
-        } catch (error) {
-            if (generation !== pollGeneration) return;
-            if (error && error.name === "AbortError") return;
-            console.warn("Monitor system stats poll failed", error);
-            lastPollFailed = true;
-        } finally {
-            if (pollController === controller) pollController = null;
-            if (generation === pollGeneration) {
-                renderLiveBadge();
-                scheduleNextPoll();
-            }
-        }
-    }
-
-    function recheck() {
-        // Recheck aborts any in-flight poll and forces a backend cache bypass.
-        pollSystemStats(true);
-    }
-
-    function reevaluatePolling() {
-        if (pollingShouldRun()) {
-            if (!pollController && !pollTimer) pollSystemStats(false);
-            else scheduleNextPoll();
-        } else {
-            cancelScheduledPoll();
-            abortInFlight();
-            pollGeneration += 1;
-            renderLiveBadge();
-        }
-    }
-
-    function onTabChanged(tabId) {
-        panelVisible = tabId === "monitor";
-        if (panelVisible) {
-            applyHiddenCardsToDom();
-            applyCardOrderToDom();
-            updateProcessHeader();
-            scrollTerminalToBottom();
-        }
-        reevaluatePolling();
-    }
-
-    function setDocumentVisibility(visible) {
-        documentVisible = Boolean(visible);
-        reevaluatePolling();
-    }
-
-    // ════════════════════════════════════════════════════════════════════
-    // Rendering: DOM construction helpers
-    // ════════════════════════════════════════════════════════════════════
-
-    function makeEl(tagName, className, text) {
-        const el = document.createElement(tagName);
-        if (className) el.className = className;
-        if (text !== undefined) el.textContent = text;
-        return el;
-    }
-
-    function setText(el, text) {
-        const value = text === null || text === undefined ? "" : String(text);
-        // Writing only on change keeps a value the user is selecting intact:
-        // replacing the text node collapses their selection on every poll.
-        if (el && el.textContent !== value) el.textContent = value;
-    }
-
-    function makeProgressBar(label, percent) {
-        const bar = makeEl("div", "progress-bar");
-        bar.setAttribute("role", "meter");
-        bar.setAttribute("aria-label", label);
-        bar.setAttribute("aria-valuemin", "0");
-        bar.setAttribute("aria-valuemax", "100");
-        bar.appendChild(makeEl("div", "progress-fill"));
-        updateProgressBar(bar, percent, label);
-        return bar;
-    }
-
-    function updateProgressBar(bar, percent, label) {
-        if (!bar) return;
-        if (label !== undefined) bar.setAttribute("aria-label", label);
-        const clamped = clampPercent(percent);
-        const fill = bar.querySelector(".progress-fill");
-        if (fill) {
-            const width = clamped === null ? "0%" : `${clamped}%`;
-            if (fill.style.width !== width) fill.style.width = width;
-        }
-        // Unavailable meters keep their node (so focus and selection survive)
-        // but render as an empty, hidden track rather than a fake zero.
-        bar.classList.toggle("hidden", clamped === null);
-        if (clamped === null) {
-            bar.removeAttribute("aria-valuenow");
-            bar.setAttribute("aria-valuetext", "Not available");
-        } else {
-            const now = String(Math.round(clamped * 10) / 10);
-            if (bar.getAttribute("aria-valuenow") !== now) bar.setAttribute("aria-valuenow", now);
-            bar.removeAttribute("aria-valuetext");
-        }
-    }
-
-    function makeMetricRow(labelText, metricName) {
-        const row = makeEl("div", "monitor-metric-row");
-        if (metricName) row.dataset.metric = metricName;
-        row.appendChild(makeEl("span", "monitor-metric-label", labelText));
-        row.appendChild(makeEl("span", "monitor-metric-reading", ""));
-        return row;
-    }
-
-    function metricRow(card, metricName) {
-        return card && card.querySelector
-            ? card.querySelector(`.monitor-metric-row[data-metric="${metricName}"]`)
-            : null;
-    }
-
-    function updateMetricRow(row, readingText, unavailable) {
-        if (!row) return;
-        const reading = row.querySelector(".monitor-metric-reading");
-        if (!reading) return;
-        setText(reading, unavailable ? "Not available" : readingText);
-        reading.classList.toggle("monitor-not-available", Boolean(unavailable));
-    }
-
-    // Fixed probe diagnostics from the backend, shown only when a vendor probe
-    // failed or found nothing. Unknown reason values render verbatim so a
-    // newer backend stays legible instead of vanishing.
-    const PROBE_REASON_LABELS = {
-        not_found: "Tool not found",
-        timeout: "Probe timed out",
-        exit_code: "Non-zero exit",
-        parse_error: "Unparsable output",
-        no_devices: "No usable devices",
-        launch_failed: "Could not launch tool",
-    };
-
-    function appendProbeDetailRows(parent, details) {
-        if (!details || typeof details !== "object") return;
-        const reason = details.reason;
-        if (typeof reason !== "string" || !reason) return;
-        const wrap = makeEl("div", "monitor-probe-details");
-        wrap.appendChild(makeMetricRow("Reason", "probe-reason"));
-        updateMetricRow(metricRow(wrap, "probe-reason"), PROBE_REASON_LABELS[reason] || reason, false);
-        if (typeof details.executable === "string" && details.executable) {
-            const row = makeMetricRow("Tool", "probe-tool");
-            updateMetricRow(row, details.executable, false);
-            wrap.appendChild(row);
-        }
-        if (typeof details.exit_code === "number") {
-            const row = makeMetricRow("Exit code", "probe-exit-code");
-            updateMetricRow(row, String(details.exit_code), false);
-            wrap.appendChild(row);
-        }
-        if (typeof details.stderr === "string" && details.stderr) {
-            const row = makeMetricRow("Stderr", "probe-stderr");
-            updateMetricRow(row, details.stderr, false);
-            wrap.appendChild(row);
-        }
-        parent.appendChild(wrap);
-    }
-
-    function makeHideButton(key, label) {
-        const button = makeEl("button", "btn btn-sm btn-ghost monitor-hide-btn", "Hide");
-        button.type = "button";
-        button.setAttribute("aria-label", `Hide ${label} monitor`);
-        button.title = `Hide ${label}`;
-        button.addEventListener("click", () => {
-            // Resolve from the card at click time: in-place updates rewrite a
-            // card's label (a GPU index or name) long after construction.
-            const card = typeof button.closest === "function"
-                ? button.closest("[data-monitor-key]")
-                : null;
-            hideCard(key, (card && card.dataset.monitorLabel) || label || key);
-        });
-        return button;
-    }
-
-    function makeDragHandle(label) {
-        const button = makeEl("button", "btn btn-sm btn-ghost monitor-drag-handle", "⠿");
-        button.type = "button";
-        button.draggable = true;
-        button.title = "Drag to reorder; use arrow keys to move";
-        button.setAttribute("aria-label", `Move ${label} monitor; use arrow keys`);
-        return button;
-    }
-
-    function cardShell(key, label, kickerText, titleText, iconClass, iconSvg) {
-        const card = makeEl("div", "card");
-        card.dataset.monitorKey = key;
-        card.dataset.monitorLabel = label;
-
-        const header = makeEl("div", "card-header");
-        const heading = makeEl("div");
-        heading.appendChild(makeEl("div", "card-kicker", kickerText));
-        heading.appendChild(makeEl("div", "card-title", titleText));
-        header.appendChild(heading);
-
-        const tools = makeEl("div", "monitor-card-tools");
-        tools.appendChild(makeDragHandle(label));
-        tools.appendChild(makeHideButton(key, label));
-        if (iconClass && iconSvg) {
-            const iconWrap = makeEl("div", `card-icon ${iconClass}`);
-            const icon = makeEl("span", "icon icon-lg");
-            icon.innerHTML = `<svg viewBox="0 0 24 24">${iconSvg}</svg>`;
-            iconWrap.appendChild(icon);
-            tools.appendChild(iconWrap);
-        }
-        header.appendChild(tools);
-        card.appendChild(header);
-        return card;
-    }
-
-    const GPU_ICON_SVG = '<polygon points="13 2 3 14 12 14 11 22 21 10 12 10 13 2"/>';
-    const PROVIDER_LABELS = { nvidia: "NVIDIA", amd: "AMD" };
-    const GPU_MONITORING_GUIDE_URL = "https://github.com/thomas9120/LLama-GUI/blob/main/docs/gpu-monitoring.md";
-
-    function providerLabel(provider) {
-        return PROVIDER_LABELS[String(provider || "").toLowerCase()] || String(provider || "GPU");
-    }
-
-    // ════════════════════════════════════════════════════════════════════
-    // Rendering: system cards (static markup, values updated in place)
-    // ════════════════════════════════════════════════════════════════════
-
-    function setMetricCard(prefix, available, percent, subText) {
-        const valueEl = byId(`monitor-${prefix}-value`);
-        const subEl = byId(`monitor-${prefix}-sub`);
-        const barHolder = byId(`monitor-${prefix}-bar`);
-        if (valueEl) {
-            if (available && clampPercent(percent) !== null) {
-                if (valueEl.childElementCount !== 2) {
-                    valueEl.replaceChildren(
-                        makeEl("span", ""),
-                        makeEl("span", "monitor-metric-unit", "%"),
-                    );
-                }
-                setText(valueEl.children[0], formatPercentValue(percent, 1));
-            } else {
-                setText(valueEl, available ? "--" : "Not available");
-            }
-        }
-        setText(subEl, available ? (subText || "") : "Collector unavailable");
-        if (barHolder) {
-            let bar = barHolder.querySelector(".progress-bar");
-            if (!bar) {
-                bar = makeProgressBar(`${prefix} usage`, null);
-                barHolder.replaceChildren(bar);
-            }
-            updateProgressBar(bar, available ? percent : null, `${prefix} usage`);
-        }
-    }
-
-    function renderSystemMetrics(system) {
-        const data = system || {};
-        const cpu = data.cpu || {};
-        const memory = data.memory || {};
-        const disk = data.disk || {};
-
-        const interval = Number(lastSample && lastSample.interval_seconds);
-        const sampleText = Number.isFinite(interval) && interval > 0
-            ? `${interval.toFixed(1)} s sample`
-            : "";
-        const intervalText = sampleText ? ` \u00b7 ${sampleText}` : "";
-
-        setMetricCard("cpu", cpu.available === true, cpu.percent,
-            cpu.available === true && cpu.percent !== null && cpu.percent !== undefined
-                ? sampleText
-                : cpu.available === true ? `Waiting for first sample${intervalText}` : "");
-
-        const memUsed = Number(memory.used_bytes);
-        const memTotal = Number(memory.total_bytes);
-        setMetricCard("memory", memory.available === true, memory.percent,
-            memory.available === true && Number.isFinite(memUsed) && Number.isFinite(memTotal)
-                ? `${formatBytes(memUsed)} used of ${formatBytes(memTotal)}`
-                : memory.available === true ? "Waiting for first sample" : "");
-
-        const readRate = finiteNonNegativeOrNull(disk.read_bytes_per_second);
-        const writeRate = finiteNonNegativeOrNull(disk.write_bytes_per_second);
-        const hasRates = readRate !== null || writeRate !== null;
-        const waiting = !hasRates && disk.io_available === true;
-        setText(byId("monitor-disk-read"), waiting ? "--" : formatRate(readRate));
-        setText(byId("monitor-disk-write"), waiting ? "--" : formatRate(writeRate));
-        const activity = !hasRates ? waiting ? "Collecting activity…" : "Disk activity unavailable"
-            : readRate > 0 && writeRate > 0 ? "Reading and writing"
-                : readRate > 0 ? "Reading" : writeRate > 0 ? "Writing"
-                    : readRate === 0 && writeRate === 0 ? "Idle" : "Partial reading";
-        setText(byId("monitor-disk-activity"), activity);
-        const scope = String(disk.io_label || "").trim();
-        setText(byId("monitor-disk-sub"), hasRates
-            ? `${scope || "Disk I/O"}${intervalText} · Includes other applications`
-            : waiting ? `${scope ? `${scope} · ` : ""}Waiting for two samples`
-                : "The system collector could not provide disk I/O readings.");
-    }
-
-    // ════════════════════════════════════════════════════════════════════
-    // Rendering: GPU + setup + state cards (rebuilt per sample)
-    // ════════════════════════════════════════════════════════════════════
-
-    function gpuCardKey(gpu) {
-        return `gpu:${String(gpu && gpu.id || "")}`;
-    }
-
-    function gpuCardName(gpu) {
-        return String(gpu && gpu.name || "").trim() || `${providerLabel(gpu && gpu.provider)} GPU`;
-    }
-
-    function gpuCardIndex(gpu) {
-        return Number.isFinite(Number(gpu && gpu.index)) ? Number(gpu.index) : null;
-    }
-
-    function makeGpuCard(gpu) {
-        const card = cardShell(gpuCardKey(gpu), "", "", "", "icon-green", GPU_ICON_SVG);
-        const utilBlock = makeEl("div", "monitor-metric-block");
-        utilBlock.appendChild(makeMetricRow("Utilization", "utilization"));
-        utilBlock.appendChild(makeProgressBar("", null));
-        card.appendChild(utilBlock);
-
-        const memBlock = makeEl("div", "monitor-metric-block");
-        memBlock.appendChild(makeMetricRow("VRAM", "vram"));
-        memBlock.appendChild(makeProgressBar("", null));
-        card.appendChild(memBlock);
-
-        const meta = makeEl("div", "monitor-gpu-meta");
-        meta.appendChild(makeMetricRow("Temperature", "temperature"));
-        meta.appendChild(makeMetricRow("GPU ID", "gpu-id"));
-        card.appendChild(meta);
-
-        updateGpuCard(card, gpu);
-        return card;
-    }
-
-    function updateGpuCard(card, gpu) {
-        const id = String(gpu && gpu.id || "");
-        const index = gpuCardIndex(gpu);
-        const name = gpuCardName(gpu);
-        const provider = providerLabel(gpu && gpu.provider);
-        const label = index === null ? name : `GPU ${index} \u00b7 ${name}`;
-        const kicker = index === null ? provider : `GPU ${index} \u00b7 ${provider}`;
-
-        card.dataset.monitorLabel = label;
-        setText(card.querySelector(".card-kicker"), kicker);
-        setText(card.querySelector(".card-title"), name);
-        const hideBtn = card.querySelector(".monitor-hide-btn");
-        if (hideBtn) {
-            hideBtn.setAttribute("aria-label", `Hide ${label} monitor`);
-            hideBtn.title = `Hide ${label}`;
-        }
-        const dragHandle = card.querySelector(".monitor-drag-handle");
-        if (dragHandle) {
-            dragHandle.setAttribute("aria-label", `Move ${label} monitor; use arrow keys`);
-            dragHandle.title = `Drag ${label} to reorder; use arrow keys to move`;
-        }
-
-        const util = gpu && gpu.utilization_percent;
-        const hasUtil = util !== null && util !== undefined;
-        updateMetricRow(metricRow(card, "utilization"),
-            hasUtil ? `${formatPercentValue(util)}%` : "", !hasUtil);
-
-        const memUsed = gpu && gpu.memory_used_bytes;
-        const memTotal = gpu && gpu.memory_total_bytes;
-        const hasMemory = memUsed !== null && memUsed !== undefined
-            && memTotal !== null && memTotal !== undefined && Number(memTotal) > 0;
-        updateMetricRow(metricRow(card, "vram"),
-            hasMemory ? `${formatBytes(memUsed)} / ${formatBytes(memTotal)}` : "", !hasMemory);
-
-        const temp = gpu && gpu.temperature_c;
-        const hasTemp = temp !== null && temp !== undefined;
-        updateMetricRow(metricRow(card, "temperature"),
-            hasTemp ? `${Math.round(Number(temp))} \u00b0C` : "", !hasTemp);
-
-        updateMetricRow(metricRow(card, "gpu-id"), shortGpuId(id), !id);
-        const idReading = metricRow(card, "gpu-id");
-        if (idReading) {
-            const reading = idReading.querySelector(".monitor-metric-reading");
-            if (reading) reading.title = id;
-        }
-
-        const bars = card.querySelectorAll(".progress-bar");
-        updateProgressBar(bars[0], hasUtil ? util : null, `${label} utilization`);
-        updateProgressBar(bars[1],
-            hasMemory ? Number(memUsed) / Number(memTotal) * 100 : null,
-            `${label} memory`);
-    }
-
-    function makeStateCard(entry) {
-        const provider = String(entry && entry.provider || "");
-        const unsupported = entry && entry.state === "unsupported";
-        const key = provider ? `state:${provider}` : "state:generic";
-        const title = provider
-            ? unsupported
-                ? `${providerLabel(provider)} SMI monitoring unavailable`
-                : `${providerLabel(provider)} telemetry unavailable`
-            : "No supported GPU telemetry detected";
-        const card = makeEl("div", "card");
-        card.dataset.monitorKey = key;
-        card.dataset.monitorLabel = title;
-
-        const tools = makeEl("div", "monitor-card-tools");
-        tools.style.justifyContent = "flex-end";
-        tools.appendChild(makeDragHandle(title));
-        tools.appendChild(makeHideButton(key, title));
-        card.appendChild(tools);
-
-        const empty = makeEl("div", "empty-state");
-        empty.appendChild(makeEl("div", "empty-state-title", title));
-        const message = String(entry && entry.message || "").trim()
-            || (provider
-                ? `System metrics keep updating. Use the ${providerLabel(provider)} setup card and Recheck when ready.`
-                : "No supported vendor tool or GPU backend identified NVIDIA or AMD hardware. System metrics keep updating; Recheck after changing the installed backend or driver environment.");
-        empty.appendChild(makeEl("p", "", message));
-        appendProbeDetailRows(empty, entry && entry.details);
-        const guide = makeEl("a", "btn btn-sm monitor-setup-guide", "GPU monitoring setup guide");
-        guide.href = GPU_MONITORING_GUIDE_URL;
-        guide.target = "_blank";
-        guide.rel = "noopener noreferrer";
-        empty.appendChild(guide);
-        card.appendChild(empty);
-        return card;
-    }
-
-    function makeSetupCard(entry) {
-        const provider = String(entry && entry.provider || "");
-        const key = `setup:${provider}`;
-        const label = `${providerLabel(provider)} monitoring setup`;
-        const card = cardShell(
-            key,
-            label,
-            "GPU Monitoring Setup",
-            providerLabel(provider),
-            provider === "amd" ? "icon-magenta" : "icon-blue",
-            GPU_ICON_SVG,
-        );
-
-        if (entry.state === "error") {
-            card.appendChild(makeEl("p", "help-text", String(entry.message || "The vendor probe failed. Check the vendor runtime installation, then Recheck.")));
-        } else if (provider === "nvidia") {
-            card.appendChild(makeEl("p", "help-text", String(entry.message || "nvidia-smi was not found. It ships with the NVIDIA driver environment \u2014 Llama GUI does not install or upgrade drivers. Install or update the driver using the official documentation, then Recheck.")));
-        } else {
-            if (entry.package_manager) {
-                const detected = makeMetricRow("Detected", "package-manager");
-                updateMetricRow(detected, `Linux \u00b7 ${entry.package_manager}`, false);
-                card.appendChild(detected);
-            }
-            card.appendChild(makeEl("p", "help-text", String(entry.message || "amd-smi was not found. The AMD repository and a compatible amdgpu driver must already be configured; then install AMD SMI once with the command shown. Llama GUI shows the command but never runs it.")));
-            if (entry.command) {
-                const row = makeEl("div", "monitor-command-row");
-                const command = makeEl("code", "monitor-command", entry.command);
-                row.appendChild(command);
-                const copyBtn = makeEl("button", "btn btn-sm", "Copy");
-                copyBtn.type = "button";
-                copyBtn.title = "Copy install command";
-                copyBtn.addEventListener("click", () => {
-                    const copyText = deps.copyText;
-                    const notify = (message, type) => {
-                        if (typeof deps.showToast === "function") deps.showToast(message, type);
-                    };
-                    if (typeof copyText !== "function") {
-                        notify("Could not copy command", "error");
-                        return;
-                    }
-                    // copyText resolves to a success flag; only claim a copy
-                    // that actually reached the clipboard.
-                    Promise.resolve(copyText(entry.command)).then(
-                        (copied) => notify(
-                            copied ? "Command copied" : "Could not copy command",
-                            copied ? "info" : "error",
-                        ),
-                        (error) => {
-                            console.warn("Could not copy install command", error);
-                            notify("Could not copy command", "error");
-                        },
-                    );
-                });
-                row.appendChild(copyBtn);
-                card.appendChild(row);
-            }
-        }
-
-        if (entry.docs_url) {
-            const actions = makeEl("div", "form-row");
-            actions.style.marginBottom = "0";
-            const link = makeEl("a", "btn", provider === "nvidia"
-                ? "Open NVIDIA driver documentation"
-                : "Open AMD SMI installation docs");
-            link.href = entry.docs_url;
-            link.target = "_blank";
-            link.rel = "noopener noreferrer";
-            actions.appendChild(link);
-            card.appendChild(actions);
-        }
-        appendProbeDetailRows(card, entry.details);
-        return card;
-    }
-
-    function stateCardKey(entry) {
-        const provider = String(entry && entry.provider || "");
-        return provider ? `state:${provider}` : "state:generic";
-    }
-
-    function setupCardKey(entry) {
-        return `setup:${String(entry && entry.provider || "")}`;
-    }
-
-    // State and setup cards are pure text plus their own controls, so a card
-    // can be reused untouched whenever the backend repeats the same advice.
-    function stateCardSignature(entry) {
-        return JSON.stringify([
-            stateCardKey(entry),
-            String(entry && entry.state || ""),
-            String(entry && entry.message || ""),
-            (entry && entry.details) || null,
-        ]);
-    }
-
-    function setupCardSignature(entry) {
-        return JSON.stringify([
-            String(entry && entry.state || ""),
-            String(entry && entry.package_manager || ""),
-            String(entry && entry.command || ""),
-            String(entry && entry.docs_url || ""),
-            String(entry && entry.message || ""),
-            (entry && entry.details) || null,
-        ]);
-    }
-
-    // Rebuilt cards are reconciled by key so a periodic refresh updates values
-    // in place. A wholesale replaceChildren() would discard the user's focus
-    // (the Hide/Copy buttons and setup links) and any text selection every
-    // two seconds, and would restart every progress-bar animation.
-    // `signatureOf` returning undefined means "always reusable"; when it
-    // returns a string, a card whose signature changed is rebuilt so its
-    // structure and listeners match the new payload.
-    function reconcileCards(container, entries, handlers) {
-        if (!container) return;
-        const { keyOf, signatureOf, create, update, owns = () => true } = handlers;
-        const existing = new Map();
-        for (const child of Array.from(container.children)) {
-            if (!owns(child)) continue;
-            const key = child && child.dataset ? child.dataset.monitorKey : null;
-            if (key && !existing.has(key)) existing.set(key, child);
-        }
-        const wanted = new Set();
-        const desired = [];
-        for (const entry of entries) {
-            const key = keyOf(entry);
-            wanted.add(key);
-            const signature = signatureOf ? signatureOf(entry) : undefined;
-            let card = existing.get(key);
-            if (card && signature !== undefined && card.dataset.monitorSig !== signature) {
-                card.remove();
-                card = null;
-            }
-            if (!card) {
-                card = create(entry);
-                card.dataset.monitorKey = key;
-                if (signature !== undefined) card.dataset.monitorSig = signature;
-                container.appendChild(card);
-            } else if (update) {
-                update(card, entry);
-            }
-            desired.push(card);
-        }
-        for (const [key, card] of existing) {
-            if (!wanted.has(key)) card.remove();
-        }
-        // Reorder only where it differs. Re-appending a card removes and
-        // re-inserts it, which drops focus and collapses a selection even
-        // though the node itself survives.
-        const current = Array.from(container.children).filter(owns);
-        let inOrder = current.length === desired.length;
-        for (let i = 0; inOrder && i < desired.length; i += 1) {
-            if (current[i] !== desired[i]) inOrder = false;
-        }
-        if (!inOrder) {
-            for (const card of desired) container.appendChild(card);
-        }
-    }
-
-    function renderGpuArea(sample) {
-        const cardGrid = byId("monitor-card-grid");
-        const stateWrap = byId("monitor-gpu-states");
-        const setupSection = byId("monitor-gpu-setup");
-        const setupCards = byId("monitor-setup-cards");
-        if (!cardGrid || !stateWrap || !setupSection || !setupCards) return;
-
-        const gpus = Array.isArray(sample && sample.gpus) ? sample.gpus : [];
-        const setupEntries = Array.isArray(sample && sample.gpu_setup) ? sample.gpu_setup : [];
-
-        reconcileCards(cardGrid, gpus, {
-            keyOf: gpuCardKey,
-            create: makeGpuCard,
-            update: updateGpuCard,
-            owns: card => String(card && card.dataset && card.dataset.monitorKey || "").startsWith("gpu:"),
-        });
-
-        // State cards: one per provider whose probe is not working, or the
-        // backend's platform-specific generic state. Keep a generic fallback
-        // for older backends that do not send that state yet.
-        const stateEntries = setupEntries.slice();
-        if (gpus.length === 0 && setupEntries.length === 0) stateEntries.push(null);
-        reconcileCards(stateWrap, stateEntries, {
-            keyOf: stateCardKey,
-            signatureOf: stateCardSignature,
-            create: makeStateCard,
-        });
-
-        // Setup cards exist only for setup_required/error states.
-        const actionable = setupEntries.filter(entry =>
-            entry.state === "setup_required" || entry.state === "error");
-        reconcileCards(setupCards, actionable, {
-            keyOf: setupCardKey,
-            signatureOf: setupCardSignature,
-            create: makeSetupCard,
-        });
-        setupSection.classList.toggle("hidden", actionable.length === 0);
-
-        setText(byId("monitor-gpu-summary"), gpus.length
-            ? stateEntries.length ? "Some GPU telemetry is unavailable" : `GPU telemetry · ${gpus.length} ${gpus.length === 1 ? "device" : "devices"}`
-            : "GPU telemetry is unavailable");
-
-        applyHiddenCardsToDom();
-        applyCardOrderToDom();
-    }
-
-    function renderSample(sample) {
-        if (!sample) return;
-        if (dragActive) {
-            // A periodic rebuild would destroy the node being dragged and
-            // cancel the drag; keep it fresh and render after the drag ends.
-            deferredRender = sample;
-            return;
-        }
-        renderSystemMetrics(sample.system);
-        renderGpuArea(sample);
-    }
-
-    // ════════════════════════════════════════════════════════════════════
-    // Rendering: process output card header
-    // ════════════════════════════════════════════════════════════════════
-
-    function renderRuntime() {
-        const state = deps.getLifecycleSnapshot?.() || {};
-        const runtime = state.activeRuntime;
-        const target = deps.getLatestStatus?.()?.external_chat_target;
-        const external = !runtime && state.phase === "idle" && target?.connected;
-        const phases = { idle: "Stopped", starting: "Starting", loading: "Loading model", ready: "Ready", running: "Running", stopping: "Stopping", failed: "Action failed" };
-        const phaseLabel = external ? "External server"
-            : runtime && state.phase === "failed" ? "Process active · action failed"
-                : phases[state.phase] || "Checking…";
-        const badge = byId("monitor-runtime-state");
-        setText(badge, phaseLabel);
-        if (badge) {
-            badge.classList.toggle("badge-green", state.phase === "ready" || state.phase === "running");
-            badge.classList.toggle("badge-yellow", Boolean(state.busy) || state.phase === "failed");
-        }
-        const model = runtime ? String(runtime.alias || runtime.model || "Model unavailable")
-            : external ? String(target.label || "Managed outside Llama GUI") : "No local process running";
-        const modelEl = byId("monitor-runtime-model");
-        setText(modelEl, runtime && !runtime.alias ? model.split(/[\\/]/).pop() : model);
-        if (modelEl) modelEl.title = model;
-        setText(byId("monitor-runtime-build"), runtime
-            ? [runtime.tool, runtime.backend, runtime.version].filter(Boolean).join(" · ")
-            : external ? "llama-server · Managed outside Llama GUI" : "Launch a model to follow its output and activity here.");
-        const endpoint = runtime?.tool === "llama-server" ? runtime : external ? target : null;
-        setText(byId("monitor-runtime-endpoint"), endpoint?.host && endpoint?.port
-            ? `Endpoint: ${endpoint.host}:${endpoint.port}` : "");
-
-        const comparison = deps.compareLaunchSettings?.(runtime);
-        const count = comparison?.available ? comparison.changes.length : 0;
-        const modelChanged = comparison?.available && (comparison.modelChanged || comparison.modelRootChanged);
-        const review = byId("btn-monitor-review");
-        if (review) {
-            review.classList.toggle("hidden", !count && !modelChanged);
-            setText(review, count ? `Review changes · ${count}` : "Review model change");
-        }
-        const note = count || modelChanged
-            ? "Edits are pending for the next launch. The count covers recorded GUI settings; API keys and Custom Launch Args are excluded."
-            : "";
-        setText(byId("monitor-runtime-note"), note);
-        byId("monitor-runtime-note")?.classList.toggle("hidden", !note);
-        setText(byId("monitor-runtime-error"), state.error || "");
-        byId("monitor-runtime-error")?.classList.toggle("hidden", !state.error);
-        byId("btn-monitor-quick-launch")?.classList.toggle("hidden", Boolean(runtime) || Boolean(state.busy) || Boolean(external));
-        byId("btn-monitor-api")?.classList.toggle("hidden", !external);
-    }
-
-    function updateProcessHeader() {
-        renderRuntime();
-        const toolBadge = byId("monitor-process-tool");
-        const stateBadge = byId("monitor-process-state");
-        const externalNote = byId("monitor-external-note");
-        const noProcessNote = byId("monitor-no-process-note");
-        const terminal = byId("output-terminal");
-
-        let lifecycle = null;
-        let status = null;
-        if (typeof deps.getLifecycleSnapshot === "function") lifecycle = deps.getLifecycleSnapshot();
-        if (typeof deps.getLatestStatus === "function") status = deps.getLatestStatus();
-
-        const runtime = lifecycle && lifecycle.activeRuntime;
-        const phase = lifecycle && lifecycle.phase;
-        const phaseLabels = { starting: "Starting", loading: "Loading", stopping: "Stopping" };
-        const transitional = Object.prototype.hasOwnProperty.call(phaseLabels, phase);
-        const running = Boolean(runtime) || Boolean(transitional);
-        const tool = runtime && runtime.tool ? runtime.tool : "";
-        const externalTarget = status && status.external_chat_target;
-        const externalConnected = phase === "idle" && Boolean(externalTarget && externalTarget.connected);
-        const navLive = byId("monitor-nav-live");
-        if (navLive) navLive.classList.toggle("hidden",
-            !((tool === "llama-server" && phase === "ready") || (!running && externalConnected)));
-
-        if (toolBadge) {
-            if (running && tool) {
-                toolBadge.textContent = tool;
-                toolBadge.classList.remove("hidden");
-                toolBadge.classList.remove("badge-accent");
-                toolBadge.classList.add("badge-neutral");
-            } else if (!running && externalConnected) {
-                toolBadge.textContent = "external server";
-                toolBadge.classList.remove("hidden");
-                toolBadge.classList.remove("badge-neutral");
-                toolBadge.classList.add("badge-accent");
-            } else {
-                toolBadge.classList.add("hidden");
-            }
-        }
-        if (stateBadge) {
-            if (running) {
-                stateBadge.textContent = phaseLabels[phase] || (phase === "failed" ? "Action failed" : phase === "ready" ? "Ready" : "Running");
-                stateBadge.classList.remove("hidden", "badge-dim", "badge-green", "badge-yellow");
-                stateBadge.classList.add(transitional || phase === "failed" ? "badge-yellow" : "badge-green");
-            } else {
-                stateBadge.textContent = "No process running";
-                stateBadge.classList.remove("hidden", "badge-green", "badge-yellow");
-                stateBadge.classList.add("badge-dim");
-            }
-        }
-        const externalOnly = externalConnected && !running;
-        const hasOutput = Boolean(terminal?.children.length);
-        setText(byId("monitor-output-title"), !running && !externalOnly && hasOutput ? "Last run output" : "Process Output");
-        setText(noProcessNote, hasOutput
-            ? "No process running — the most recent output backlog is retained until the next launch."
-            : "Process output will appear here when you launch a model.");
-        if (externalNote) externalNote.classList.toggle("hidden", !externalOnly);
-        if (noProcessNote) noProcessNote.classList.toggle("hidden", running || externalOnly);
-        if (terminal) terminal.classList.toggle("hidden", externalOnly || (!running && !hasOutput));
-        renderInferenceAvailability(deps.getInferenceSnapshot?.());
-        // #input-row visibility belongs to app.js, which owns process
-        // lifecycle and sets it from the active runtime before calling
-        // updateProcessHeader(); touching it here would fight that.
-    }
-
-    // ════════════════════════════════════════════════════════════════════
-    // Process output terminal (rendering only; cursor stays in app.js)
-    // ════════════════════════════════════════════════════════════════════
-
-    function terminalEl() {
-        return byId("output-terminal");
-    }
-
-    function scrollTerminalToBottom() {
-        const terminal = terminalEl();
-        if (terminal) terminal.scrollTop = terminal.scrollHeight;
-    }
-
-    function trimTerminal() {
-        const terminal = terminalEl();
-        if (!terminal) return;
-        if (terminal.childElementCount <= TERMINAL_MAX_LINES) return;
-        const range = document.createRange ? document.createRange() : null;
-        if (range) {
-            range.setStartBefore(terminal.firstElementChild);
-            range.setEndAfter(terminal.children[TERMINAL_TRIM - 1]);
-            range.deleteContents();
-        } else {
-            for (let i = 0; i < TERMINAL_TRIM; i += 1) {
-                if (terminal.firstElementChild) terminal.firstElementChild.remove();
-            }
-        }
-    }
-
-    function appendOutputLine(text) {
-        const terminal = terminalEl();
-        if (!terminal) return;
-        const line = document.createElement("div");
-        line.textContent = text;
-        terminal.appendChild(line);
-        if (terminal.children.length === 1) updateProcessHeader();
-        trimTerminal();
-        scrollTerminalToBottom();
-    }
-
-    function clearTerminal() {
-        const terminal = terminalEl();
-        if (terminal) terminal.replaceChildren();
-        // Advance the cursor epoch without discarding the cursor; otherwise
-        // the next cursorless request would replay the whole backend backlog.
-        if (typeof deps.invalidateCursor === "function") deps.invalidateCursor();
-        updateProcessHeader();
-    }
-
-    // ════════════════════════════════════════════════════════════════════
-    // Inference card (snapshots supplied by app.js's shared poller)
-    // ════════════════════════════════════════════════════════════════════
-
-    function setInferenceText(id, text) {
-        const el = byId(id);
-        if (el) el.textContent = text;
-    }
-
-    function renderInferenceAvailability(snapshot) {
-        const state = deps.getLifecycleSnapshot?.() || {};
-        let message = "";
-        if (!snapshot?.targetKey) {
-            message = state.phase === "starting" || state.phase === "loading"
-                ? "Waiting for llama-server to become ready. Follow loading progress in Process Output."
-                : state.phase === "stopping"
-                    ? "The process is stopping. Inference readings will resume when a server is ready."
-                    : state.activeRuntime && state.activeRuntime.tool !== "llama-server"
-                        ? `${state.activeRuntime.tool} does not expose server inference metrics. Follow its output above; system readings remain available independently.`
-                        : "Start or connect to a llama-server to view token, request, and context activity. GPU monitoring tools are not required.";
-        } else if (snapshot.seq === 1) {
-            message = "Waiting for the first inference sample from this server.";
-        } else {
-            const notes = [];
-            if (snapshot.sources?.metrics !== "ok") notes.push("Session counters unavailable: check the server connection and whether --metrics is enabled.");
-            if (snapshot.sources?.slots !== "ok") notes.push("Slot activity and context unavailable: check the server connection and whether /slots is enabled.");
-            message = notes.join(" ");
-        }
-        setText(byId("monitor-inference-note"), message);
-        byId("monitor-inference-note")?.classList.toggle("hidden", !message);
-    }
-
-    function renderInferenceSnapshot(snapshot) {
-        renderInferenceAvailability(snapshot);
-        const kicker = byId("monitor-inference-kicker");
-        const badge = byId("monitor-inference-state-badge");
-        const body = byId("monitor-inference-body");
-        const empty = byId("monitor-inference-empty");
-        if (!kicker || !badge || !body || !empty) return;
-
-        if (!snapshot || !snapshot.targetKey) {
-            kicker.textContent = "Llama server";
-            badge.textContent = "Unavailable";
-            badge.classList.remove("badge-green", "badge-neutral");
-            badge.classList.add("badge-dim");
-            badge.classList.remove("hidden");
-            body.classList.add("hidden");
-            empty.classList.remove("hidden");
-            return;
-        }
-
-        const processing = snapshot.requests ? snapshot.requests.processingBest : null;
-        const activityUnknown = processing === null || processing === undefined;
-        const stateText = activityUnknown ? "Activity unknown" : processing > 0 ? "Processing" : "Idle";
-        const external = String(snapshot.targetKey).startsWith("ext:");
-        const target = external ? deps.getLatestStatus?.()?.external_chat_target : null;
-        kicker.textContent = external && target?.host && target?.port
-            ? `External llama-server · ${target.host}:${target.port}`
-            : `Llama server · ${stateText}`;
-        if (processing !== null && processing > 0) {
-            badge.textContent = `${processing} active`;
-            badge.classList.remove("badge-dim", "badge-neutral");
-            badge.classList.add("badge-green");
-        } else if (activityUnknown) {
-            badge.textContent = stateText;
-            badge.classList.remove("badge-green", "badge-neutral");
-            badge.classList.add("badge-dim");
-        } else {
-            badge.textContent = stateText;
-            badge.classList.remove("badge-green", "badge-dim");
-            badge.classList.add("badge-neutral");
-        }
-        badge.classList.remove("hidden");
-        empty.classList.add("hidden");
-        body.classList.remove("hidden");
-
-        const session = snapshot.session || {};
-        setInferenceText("monitor-inference-prompt",
-            session.prompt === null ? "--" : `${formatTokens(session.prompt)} tokens`);
-        setInferenceText("monitor-inference-generated",
-            session.generated === null ? "--" : `${formatTokens(session.generated)} tokens`);
-        setInferenceText("monitor-inference-total",
-            session.total === null ? "--" : `${formatTokens(session.total)} tokens`);
-
-        const context = snapshot.context;
-        const contextLabel = byId("monitor-inference-context-label");
-        const contextReading = byId("monitor-inference-context-reading");
-        const contextBarHolder = byId("monitor-inference-context-bar");
-        if (contextLabel) {
-            contextLabel.textContent = context
-                ? `Context \u00b7 most-filled slot (${context.isProcessing ? "active" : "idle"})`
-                : "Context \u00b7 most-filled slot";
-        }
-        if (contextReading) {
-            contextReading.textContent = context
-                ? `${formatTokens(context.used)} / ${formatTokens(context.total)} \u00b7 ${formatTokens(context.remaining)} remaining`
-                : "Not available";
-            contextReading.classList.toggle("monitor-not-available", !context);
-            contextReading.classList.toggle("monitor-metric-reading", Boolean(context));
-        }
-        if (contextBarHolder) {
-            let bar = contextBarHolder.querySelector(".progress-bar");
-            if (!bar) {
-                bar = makeProgressBar("Most-filled slot context usage", null);
-                contextBarHolder.replaceChildren(bar);
-            }
-            updateProgressBar(bar, context ? context.percent : null, "Most-filled slot context usage");
-            const fill = bar.querySelector(".progress-fill");
-            if (fill) {
-                fill.classList.toggle("progress-fill-critical", Boolean(context) && snapshot.contextLevel === "critical");
-                fill.classList.toggle("progress-fill-warning", Boolean(context) && snapshot.contextLevel === "warning");
-            }
-        }
-
-        const speed = snapshot.speed || {};
-        setInferenceText("monitor-inference-prompt-speed-label",
-            speed.promptIsLive ? "Live prompt speed" : "Avg prompt speed");
-        setInferenceText("monitor-inference-gen-speed-label",
-            speed.generatedIsLive ? "Live generation speed" : "Avg generation speed");
-        setInferenceText("monitor-inference-prompt-speed",
-            speed.prompt === null || speed.prompt === undefined ? "--" : `${formatSpeed(speed.prompt)} tok/s`);
-        setInferenceText("monitor-inference-gen-speed",
-            speed.generated === null || speed.generated === undefined ? "--" : `${formatSpeed(speed.generated)} tok/s`);
-
-        const requests = snapshot.requests || {};
-        const requestParts = [];
-        if (requests.processing !== null && requests.processing !== undefined) {
-            requestParts.push(`${requests.processing} active`);
-        }
-        if (requests.queued !== null && requests.queued !== undefined) {
-            requestParts.push(`${requests.queued} queued`);
-        }
-        setInferenceText("monitor-inference-requests", requestParts.length ? requestParts.join(" \u00b7 ") : "--");
-
-        setInferenceText("monitor-inference-slots", snapshot.slots
-            ? `${snapshot.slots.busy} / ${snapshot.slots.total} busy`
-            : "--");
-    }
-
-    // The fixed stats bar is part of the shared Chat shell, while its values
-    // come from the one main-window inference engine. Keep this renderer pure
-    // so a detached Chat can reuse the DOM helper without creating a poller.
-    function renderStatsBarFromSnapshot(snapshot, targetDocument = document) {
-        const doc = targetDocument || document;
-        const get = id => doc.getElementById(id);
-        const set = (id, value, format) => {
-            const element = get(id);
-            if (element) element.textContent = value === null || value === undefined ? "--" : format(value);
-        };
-        const bar = get("stats-bar");
-        if (!bar) return;
-        if (!snapshot || !snapshot.targetKey) {
-            bar.classList.add("hidden");
-            for (const id of ["stats-prompt-tokens", "stats-prompt-speed", "stats-gen-tokens", "stats-gen-speed", "stats-context"]) {
-                const element = get(id);
-                if (element) element.textContent = "--";
-            }
-            const kv = get("stats-kv-usage");
-            if (kv) kv.textContent = "--%";
-            return;
-        }
-        bar.classList.remove("hidden");
-        set("stats-prompt-tokens", snapshot.session?.prompt, value => Math.round(value).toLocaleString());
-        set("stats-prompt-speed", snapshot.speed?.prompt, value => value.toFixed(1));
-        set("stats-prompt-speed-label", snapshot.speed?.promptIsLive, live => live ? "tok/s prompt live" : "tok/s prompt avg");
-        set("stats-gen-tokens", snapshot.session?.generated, value => Math.round(value).toLocaleString());
-        set("stats-gen-speed", snapshot.speed?.generated, value => value.toFixed(1));
-        set("stats-gen-speed-label", snapshot.speed?.generatedIsLive, live => live ? "tok/s gen live" : "tok/s gen avg");
-        set("stats-context", snapshot.session?.total, value => Math.round(value).toLocaleString());
-        set("stats-kv-usage", snapshot.context ? snapshot.context.percent : null, value => `${Math.round(value)}%`);
-    }
-
-    // ════════════════════════════════════════════════════════════════════
-    // Hidden-card behavior
-    // ════════════════════════════════════════════════════════════════════
-
-    function loadHiddenCards() {
-        hiddenCards = new Map();
-        try {
-            const raw = localStorage.getItem(HIDDEN_STORAGE_KEY);
-            if (!raw) return;
-            for (const entry of normalizeHiddenEntries(JSON.parse(raw))) {
-                hiddenCards.set(entry.key, { label: entry.label, sessionOnly: isSessionOnlyKey(entry.key) });
-            }
-        } catch (error) {
-            console.debug("Could not read hidden monitor cards", error);
-        }
-    }
-
-    function persistHiddenCards() {
-        try {
-            const entries = [];
-            for (const [key, value] of hiddenCards) {
-                if (value.sessionOnly) continue;
-                entries.push({ key, label: value.label });
-            }
-            const retained = entries.slice(-HIDDEN_MAX_ENTRIES);
-            localStorage.setItem(HIDDEN_STORAGE_KEY, JSON.stringify(retained));
-            if (entries.length > HIDDEN_MAX_ENTRIES) {
-                const retainedKeys = new Set(retained.map(entry => entry.key));
-                for (const [key, value] of hiddenCards) {
-                    if (!value.sessionOnly && !retainedKeys.has(key)) hiddenCards.delete(key);
-                }
-            }
-        } catch (error) {
-            console.debug("Could not persist hidden monitor cards", error);
-        }
-    }
-
-    function hiddenBarEls() {
-        return {
-            controls: byId("monitor-hidden-controls"),
-            count: byId("monitor-hidden-count"),
-            items: byId("monitor-restore-items"),
-        };
-    }
-
-    function applyHiddenCardsToDom() {
-        const cards = Array.from(document.querySelectorAll("[data-monitor-key]"));
-        const relevant = [];
-        for (const card of cards) {
-            const key = card.dataset.monitorKey;
-            const hidden = hiddenCards.has(key);
-            card.classList.toggle("hidden", hidden);
-            if (hidden) relevant.push(card);
-        }
-
-        const { controls, count, items } = hiddenBarEls();
-        if (!controls || !count || !items) return;
-        controls.classList.toggle("hidden", relevant.length === 0);
-        if (relevant.length === 0) controls.open = false;
-        count.textContent = relevant.length === 1 ? "1 card hidden" : `${relevant.length} cards hidden`;
-
-        const signature = relevant.map(card => `${card.dataset.monitorKey}\u0000${card.dataset.monitorLabel || ""}`).join("\u0001");
-        if (signature === hiddenRestoreSignature) return;
-        hiddenRestoreSignature = signature;
-        items.replaceChildren();
-        for (const card of relevant) {
-            const key = card.dataset.monitorKey;
-            const label = card.dataset.monitorLabel || key;
-            const row = makeEl("div", "monitor-restore-row");
-            row.appendChild(makeEl("span", "", label));
-            const restore = makeEl("button", "btn btn-sm btn-ghost", "Show");
-            restore.type = "button";
-            restore.setAttribute("aria-label", `Show ${label} monitor`);
-            restore.addEventListener("click", () => restoreCard(key, card));
-            row.appendChild(restore);
-            items.appendChild(row);
-        }
-    }
-
-    function hideCard(key, label) {
-        if (!key) return;
-        hiddenCards.delete(key);
-        hiddenCards.set(key, { label: label || key, sessionOnly: isSessionOnlyKey(key) });
-        persistHiddenCards();
-        applyHiddenCardsToDom();
-        // Hiding the focused card moves focus to the restore control.
-        const { controls } = hiddenBarEls();
-        if (controls) {
-            const summary = controls.querySelector("summary");
-            if (summary && typeof summary.focus === "function") summary.focus();
-        }
-    }
-
-    function restoreCard(key, cardEl) {
-        hiddenCards.delete(key);
-        persistHiddenCards();
-        applyHiddenCardsToDom();
-        const card = cardEl || document.querySelector(`[data-monitor-key="${CSS.escape ? CSS.escape(key) : key}"]`);
-        if (card) {
-            const disclosure = card.closest?.("#monitor-gpu-help");
-            if (disclosure) disclosure.open = true;
-            const target = card.querySelector(".card-title") || card.querySelector(".monitor-hide-btn") || card;
-            if (target && typeof target.focus === "function") {
-                if (!target.hasAttribute || !target.hasAttribute("tabindex")) {
-                    if (typeof target.setAttribute === "function") target.setAttribute("tabindex", "-1");
-                }
-                target.focus();
-            }
-        }
-    }
-
-    function showAllCards() {
-        const disclosure = byId("monitor-gpu-help");
-        if (disclosure?.querySelector(".card.hidden")) disclosure.open = true;
-        hiddenCards = new Map();
-        persistHiddenCards();
-        applyHiddenCardsToDom();
-    }
-
-    // ════════════════════════════════════════════════════════════════════
-    // Card ordering (drag and drop, per container)
-    // ════════════════════════════════════════════════════════════════════
-
-    function loadCardOrder() {
-        cardOrder = [];
-        try {
-            const raw = localStorage.getItem(ORDER_STORAGE_KEY);
-            if (raw) cardOrder = normalizeOrderEntries(JSON.parse(raw));
-        } catch (error) {
-            console.debug("Could not read monitor card order", error);
-        }
-    }
-
-    function persistCardOrder() {
-        try {
-            // Index-fallback GPU keys are session-only, exactly like hides.
-            const retained = normalizeOrderEntries(
-                cardOrder.filter(key => !isSessionOnlyKey(key))
-            );
-            localStorage.setItem(ORDER_STORAGE_KEY, JSON.stringify(retained));
-        } catch (error) {
-            console.debug("Could not persist monitor card order", error);
-        }
-    }
-
-    function applyCardOrderToDom() {
-        if (typeof document.querySelectorAll !== "function") return;
-        for (const container of document.querySelectorAll(ORDER_CONTAINER_SELECTOR)) {
-            if (!container || !container.children) continue;
-            const cards = Array.from(container.children).filter(
-                child => child && child.dataset && child.dataset.monitorKey
-            );
-            if (cards.length < 2) continue;
-            // Cards absent from the persisted order keep their current DOM
-            // order after every known card.
-            const currentIndex = new Map();
-            cards.forEach((card, index) => currentIndex.set(card.dataset.monitorKey, index));
-            const orderOf = (card) => {
-                const known = cardOrder.indexOf(card.dataset.monitorKey);
-                return known !== -1 ? known : cardOrder.length + currentIndex.get(card.dataset.monitorKey);
-            };
-            const sorted = cards.slice().sort((a, b) => orderOf(a) - orderOf(b));
-            // Re-appending a card removes and re-inserts it, which drops focus
-            // and collapses a selection even though the node survives; only
-            // touch the DOM when the order is actually wrong.
-            let inOrder = true;
-            for (let i = 0; i < sorted.length; i += 1) {
-                if (sorted[i] !== cards[i]) { inOrder = false; break; }
-            }
-            if (inOrder) continue;
-            for (const card of sorted) container.appendChild(card);
-        }
-    }
-
-    function monitorCardFromEvent(event) {
-        const target = event && event.target;
-        if (!target || typeof target.closest !== "function") return null;
-        return target.closest("[data-monitor-key]");
-    }
-
-    function clearDropMarkers() {
-        if (typeof document.querySelectorAll !== "function") return;
-        for (const card of document.querySelectorAll("[data-monitor-key]")) {
-            card.classList.remove("drop-before", "drop-after", "drop-horizontal");
-        }
-    }
-
-    // Merge one container's new key order into the global order. Only relative
-    // order within the same container matters, so the container keys simply
-    // move behind the others; unknown keys stay untouched and dead keys stay
-    // inert (they can resurrect a card that reappears).
-    function updateCardOrder(containerKeys) {
-        const other = cardOrder.filter(key => !containerKeys.includes(key));
-        cardOrder = other.concat(containerKeys);
-        persistCardOrder();
-        applyCardOrderToDom();
-    }
-
-    function dropPlacement(container, dropCard, clientX, clientY) {
-        const rect = typeof dropCard.getBoundingClientRect === "function"
-            ? dropCard.getBoundingClientRect()
-            : null;
-        if (!rect) return { before: false, horizontal: false };
-        const horizontal = Array.from(container.children).some((other) => {
-            if (other === dropCard || (other.classList && other.classList.contains("hidden"))
-                    || typeof other.getBoundingClientRect !== "function") return false;
-            const otherRect = other.getBoundingClientRect();
-            return rect.width > 0 && otherRect.width > 0
-                && Math.max(rect.top, otherRect.top) < Math.min(rect.bottom, otherRect.bottom);
-        });
-        const point = Number(horizontal ? clientX : clientY);
-        const midpoint = horizontal
-            ? rect.left + rect.width / 2
-            : rect.top + rect.height / 2;
-        return {
-            before: Number.isFinite(point) ? point < midpoint : false,
-            horizontal,
-        };
-    }
-
-    function reorderCard(container, draggedKey, dropCard, clientX, clientY) {
-        const keys = Array.from(container.children)
-            .filter(child => child && child.dataset && child.dataset.monitorKey)
-            .map(child => child.dataset.monitorKey);
-        const remaining = keys.filter(key => key !== draggedKey);
-        if (remaining.length === keys.length) return;
-        let insertAt;
-        if (dropCard) {
-            const index = remaining.indexOf(dropCard.dataset.monitorKey);
-            if (index === -1) return;
-            const { before } = dropPlacement(container, dropCard, clientX, clientY);
-            insertAt = before ? index : index + 1;
-        } else {
-            // Dropped on empty container space: append at the end.
-            insertAt = remaining.length;
-        }
-        remaining.splice(insertAt, 0, draggedKey);
-        updateCardOrder(remaining);
-    }
-
-    function onKeyDown(container, event) {
-        const target = event && event.target;
-        if (!target || typeof target.closest !== "function"
-                || !target.closest(ORDER_DRAG_HANDLE_SELECTOR)) return;
-        const offset = event.key === "ArrowLeft" || event.key === "ArrowUp"
-            ? -1
-            : event.key === "ArrowRight" || event.key === "ArrowDown" ? 1 : 0;
-        if (!offset) return;
-        event.preventDefault();
-        const card = monitorCardFromEvent(event);
-        const cards = Array.from(container.children).filter(
-            child => child && child.dataset && child.dataset.monitorKey
-                && !(child.classList && child.classList.contains("hidden"))
-        );
-        const index = cards.indexOf(card);
-        const targetCard = cards[index + offset];
-        if (index === -1 || !targetCard) return;
-        const keys = Array.from(container.children)
-            .filter(child => child && child.dataset && child.dataset.monitorKey)
-            .map(child => child.dataset.monitorKey)
-            .filter(key => key !== card.dataset.monitorKey);
-        const targetIndex = keys.indexOf(targetCard.dataset.monitorKey);
-        keys.splice(offset < 0 ? targetIndex : targetIndex + 1, 0, card.dataset.monitorKey);
-        updateCardOrder(keys);
-        if (typeof target.focus === "function") target.focus();
-    }
-
-    function onDragStart(container, event) {
-        const target = event && event.target;
-        if (!target || typeof target.closest !== "function"
-                || !target.closest(ORDER_DRAG_HANDLE_SELECTOR)) return;
-        const card = monitorCardFromEvent(event);
-        if (!card) return;
-        dragKey = card.dataset.monitorKey;
-        dragContainer = container;
-        dragActive = true;
-        card.classList.add("dragging");
-        if (event.dataTransfer) {
-            event.dataTransfer.effectAllowed = "move";
-            try {
-                event.dataTransfer.setData("text/plain", dragKey);
-            } catch (error) {
-                console.debug("Could not set drag payload", error);
-            }
-        }
-    }
-
-    function onDragOver(container, event) {
-        if (!dragActive || !dragKey) return;
-        if (dragContainer !== container) {
-            clearDropMarkers();
-            return;
-        }
-        const card = monitorCardFromEvent(event);
-        if (event.dataTransfer) event.dataTransfer.dropEffect = "move";
-        clearDropMarkers();
-        if (card && card.dataset.monitorKey !== dragKey) {
-            event.preventDefault();
-            const { before, horizontal } = dropPlacement(
-                container, card, event.clientX, event.clientY
-            );
-            card.classList.toggle("drop-horizontal", horizontal);
-            card.classList.add(before ? "drop-before" : "drop-after");
-        } else if (!card) {
-            event.preventDefault();
-            // Empty container space appends at the end: mark the last card.
-            const cards = Array.from(container.children).filter(
-                child => child && child.dataset && child.dataset.monitorKey
-            );
-            const last = cards[cards.length - 1];
-            if (last) last.classList.add("drop-after");
-        }
-    }
-
-    function onDrop(container, event) {
-        if (!dragActive || !dragKey || dragContainer !== container) return;
-        event.preventDefault();
-        reorderCard(
-            container, dragKey, monitorCardFromEvent(event), event.clientX, event.clientY
-        );
-        finishDrag();
-    }
-
-    function finishDrag() {
-        if (!dragActive) return;
-        dragKey = null;
-        dragContainer = null;
-        dragActive = false;
-        if (typeof document.querySelectorAll === "function") {
-            for (const card of document.querySelectorAll("[data-monitor-key]")) {
-                card.classList.remove("dragging", "drop-before", "drop-after", "drop-horizontal");
-            }
-        }
-        if (deferredRender !== null) {
-            const sample = deferredRender;
-            deferredRender = null;
-            renderSample(sample);
-        }
-    }
-
-    // Test-only: drop module polling/preference state between scenarios.
-    function resetForTests() {
-        cancelScheduledPoll();
-        abortInFlight();
-        pollGeneration += 1;
-        panelVisible = false;
-        documentVisible = true;
-        lastSample = null;
-        lastSuccessAt = 0;
-        lastPollFailed = false;
-        everSucceeded = false;
-        hiddenCards = new Map();
-        hiddenRestoreSignature = "";
-        cardOrder = [];
-        dragKey = null;
-        dragContainer = null;
-        dragActive = false;
-        deferredRender = null;
-    }
-
-    // ════════════════════════════════════════════════════════════════════
-    // Init / wiring
-    // ════════════════════════════════════════════════════════════════════
-
-    function init() {
-        loadHiddenCards();
-        loadCardOrder();
-
-        const recheckBtn = byId("btn-monitor-recheck");
-        if (recheckBtn) recheckBtn.addEventListener("click", recheck);
-
-        for (const [id, tab] of [["btn-monitor-configure", "configure"], ["btn-monitor-quick-launch", "quick-launch"], ["btn-monitor-api", "api"]]) {
-            byId(id)?.addEventListener("click", () => deps.switchTab(tab));
-        }
-        byId("btn-monitor-review")?.addEventListener("click", () => deps.reviewLaunchChanges());
-
-        const resetBtn = byId("btn-reset-inference");
-        if (resetBtn) {
-            resetBtn.addEventListener("click", () => {
-                if (typeof deps.resetStatsBaseline === "function") deps.resetStatsBaseline();
-            });
-        }
-
-        const showAllBtn = byId("btn-monitor-show-all");
-        if (showAllBtn) showAllBtn.addEventListener("click", showAllCards);
-
-        // Static cards (system/inference) declare their hide buttons inline.
-        const staticHideButtons = document.querySelectorAll
-            ? document.querySelectorAll("[data-monitor-hide]")
-            : [];
-        for (const button of staticHideButtons) {
-            button.addEventListener("click", () => {
-                const key = button.dataset.monitorHide;
-                const card = typeof button.closest === "function"
-                    ? button.closest("[data-monitor-key]")
-                    : null;
-                const label = (card && card.dataset.monitorLabel) || key;
-                hideCard(key, label);
-            });
-        }
-
-        // Drag-and-drop reordering, delegated per container so per-sample card
-        // rebuilds need no re-binding. Cards move within their own container
-        // only (see ORDER_CONTAINER_SELECTOR).
-        if (typeof document.querySelectorAll === "function") {
-            for (const container of document.querySelectorAll(ORDER_CONTAINER_SELECTOR)) {
-                container.addEventListener("keydown", (event) => onKeyDown(container, event));
-                container.addEventListener("dragstart", (event) => onDragStart(container, event));
-                container.addEventListener("dragover", (event) => onDragOver(container, event));
-                container.addEventListener("drop", (event) => onDrop(container, event));
-                container.addEventListener("dragend", () => finishDrag());
-            }
-        }
-
-        updateProcessHeader();
-        renderLiveBadge();
-        applyHiddenCardsToDom();
-        applyCardOrderToDom();
-    }
-
-    root.monitorUi = {
-        configure,
-        init,
-        onTabChanged,
-        setDocumentVisibility,
-        recheck,
-        appendOutputLine,
-        clearTerminal,
-        updateProcessHeader,
-        renderRuntime,
-        renderInferenceSnapshot,
-        renderStatsBarFromSnapshot,
-        // Compatibility aliases; new consumers use root.inferenceStats directly.
-        createInferenceStats: root.inferenceStats.createInferenceStats,
-        parseMetricsText: root.inferenceStats.parseMetricsText,
-        normalizeSlots: root.inferenceStats.normalizeSlots,
-        // Exported for unit tests:
-        formatBytes,
-        formatRate,
-        formatPercentValue,
-        formatTokens,
-        formatClock,
-        shortGpuId,
-        normalizeHiddenEntries,
-        isSessionOnlyKey,
-        normalizeOrderEntries,
-        _resetForTests: resetForTests,
-    };
-})();
