+++ ui/js/monitor/monitor-internal.js
@@ -1,0 +1,5 @@
+// Private package links and live injected dependencies; each concern owns its state.
+(() => {
+    "use strict";
+    window.LlamaGui._monitorInternal = { dependencies: {} };
+})();
